backend/requirements.txt
pyproject-toml
psycopg2
yarl==1.8.1
Django
environs
gunicorn
pandas
djangorestframework
openpyxl
requests
django-admin-rangefilter
django-filter
celery
django-celery-beat
django-celery-results
Redis
django-redis
pillow
pytesseract
beautifulsoup4
XlsxWriter
selenium
webdriver-manager